consumer根据offset查询功能完善

+++ shinano-mq-base/src/main/java/cn/com/shinano/ShinanoMQ/base/VO/MessageListVO.java
@@ -16,5 +16,5 @@ import java.util.List;
 @NoArgsConstructor
 public class MessageListVO {
     private List<SaveMessage> messages;
-    private Long nextOffset;
+    private Integer nextOffset;
 }
